+++ packages/layout-core/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @orderly.network/layout-core
 
+## 2.0.4-alpha.3
+
+### Patch Changes
+
+- publish
+
+## 2.0.4-alpha.2
+
+### Patch Changes
+
+- publish
+
+## 2.0.4-alpha.1
+
+### Patch Changes
+
+- publish
+
+## 2.0.4-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 2.0.3
+
+### Patch Changes
+
+- ccfcda8: publish
+- publish
+
+## 2.0.3-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 2.0.2
 
 ### Patch Changes