+++ packages/layout-grid/CHANGELOG.md
@@ -1,5 +1,67 @@
 # @orderly.network/layout-grid
 
+## 2.0.4-alpha.3
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/layout-core@2.0.4-alpha.3
+  - @orderly.network/trading-next@3.0.4-alpha.3
+  - @orderly.network/ui@3.0.4-alpha.3
+
+## 2.0.4-alpha.2
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/layout-core@2.0.4-alpha.2
+  - @orderly.network/trading-next@3.0.4-alpha.2
+  - @orderly.network/ui@3.0.4-alpha.2
+
+## 2.0.4-alpha.1
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/layout-core@2.0.4-alpha.1
+  - @orderly.network/trading-next@3.0.4-alpha.1
+  - @orderly.network/ui@3.0.4-alpha.1
+
+## 2.0.4-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/layout-core@2.0.4-alpha.0
+  - @orderly.network/trading-next@3.0.4-alpha.0
+  - @orderly.network/ui@3.0.4-alpha.0
+
+## 2.0.3
+
+### Patch Changes
+
+- ccfcda8: publish
+- publish
+- Updated dependencies [ccfcda8]
+- Updated dependencies
+  - @orderly.network/layout-core@2.0.3
+  - @orderly.network/trading-next@3.0.3
+  - @orderly.network/ui@3.0.3
+
+## 2.0.3-alpha.0
+
+### Patch Changes
+
+- publish
+- Updated dependencies
+  - @orderly.network/layout-core@2.0.3-alpha.0
+  - @orderly.network/trading-next@3.0.3-alpha.0
+  - @orderly.network/ui@3.0.3-alpha.0
+
 ## 2.0.2
 
 ### Patch Changes