+++ packages/types/CHANGELOG.md
@@ -1,5 +1,42 @@
 # @orderly.network/types
 
+## 3.0.4-alpha.3
+
+### Patch Changes
+
+- publish
+
+## 3.0.4-alpha.2
+
+### Patch Changes
+
+- publish
+
+## 3.0.4-alpha.1
+
+### Patch Changes
+
+- publish
+
+## 3.0.4-alpha.0
+
+### Patch Changes
+
+- publish
+
+## 3.0.3
+
+### Patch Changes
+
+- ccfcda8: publish
+- publish
+
+## 3.0.3-alpha.0
+
+### Patch Changes
+
+- publish
+
 ## 3.0.2
 
 ### Patch Changes